requirements.txt
crewai>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.8.2
requests>=2.32.3
rapidfuzz>=3.9.6
docx==0.2.4
python-docx>=1.1.2
openai>=1.30.0
qdrant-client>=1.9.0
psycopg2-binary>=2.9.9
flask>=3.0.0
flask_sqlalchemy>=3.1.1
